Ads moved, Tag cleanup
Capitalization in tags for Red Hat
cicd replaced with CI/CD
CI/CD added where appropriate

Moved Carbon Ads out of the sidebar (ugly) to the tops of posts (after the metadata); I'm not attached to that placement, but it's vital to load the ad code and the page as fast as possible. We'll test it and see if anyone notices. Alternative position is immediatly after .Content.

Signed-off-by: Chris Short <chris@chrisshort.net>