Return of the G
In general, I'm not happy with SendInBlue's subscription form maker but, it does handle things way better than Buttondown

For whatever reason, I could not get a partial for GDPR compliance onto the subscribe page. So I inserted the HTML directly. Shame on me and it was even worse. I think SIB is trying to do too much with their forms. Drop the outer container somehow.